Update Product Info page style.
Add a fixed center layout.
